Rename package to dashboard
